Added Task to Risk Assessment

+++ app/javascript/components/NewClientForm/taskList.js
@@ -11,7 +11,9 @@ export default (props) => {
   } = props
 
   const handleRemoveTask = index => {
-    deleteTask(index)
+    let isYes = confirm("Are you sure to delete task?");
+    if (isYes)
+      deleteTask(index)
   }
 
   const onChange = (e) => {
@@ -19,34 +21,47 @@ export default (props) => {
       task.expected_date = e.data
     else
       task.name = e.target.value
-
   }
 
   return (
     <div className="row">
       <div className="col-xs-12 col-md-5 col-lg-5">
-        <TextInput
-          label={"Name"}
-          onChange={onChange}
-          value={task.name}
-        />
+        {
+          task.completed ? <div>
+            <label>Name</label>
+            <p>{task.name}</p>
+          </div>
+          :
+          <TextInput
+            label={"Name"}
+            onChange={onChange}
+            value={task.name}
+          />
+        }
+
       </div>
       <div className="col-xs-12 col-md-5 col-lg-5">
-        <DateInput
-          T={'Expected Date'}
-          isError={false}
-          label={"Expected Date"}
-          value={task.expected_date}
-          onChange={onChange}
+        { task.completed ? <div>
+            <label>Expected Date</label>
+            <p>{task.expected_date}</p>
+          </div>
+          :
+          <DateInput
+            T={'Expected Date'}
+            isError={false}
+            label={"Expected Date"}
+            value={task.expected_date}
+            onChange={onChange}
           />
+        }
       </div>
       <div className="col-xs-12 col-md-2">
-        <button
+        { !task.completed && <button
           className="btn btn-danger m-t-lg"
           onClick={(e) => { e.preventDefault(); handleRemoveTask(index) }}
         >
           <i className="fa fa-trash"></i>
-        </button>
+        </button> }
       </div>
     </div>
   )
